backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0.post1
firebase-admin==6.2.0
python-dotenv==1.0.0
pydantic==2.5.2
httpx==0.25.1
